various fixes towards a mastr scenario loader this needs https://github.com/NOWUM/open-energy-data-server/
