+++ tests/test_flexable_strategies.py
@@ -24,7 +24,7 @@ def power_plant() -> PowerPlant:
     return PowerPlant(
         id="test_pp",
         unit_operator="test_operator",
-        technology="coal",
+        technology="hard coal",
         index=index,
         max_power=1000,
         min_power=200,
